+++ src/SFA.DAS.EAS.Employer_Account.Database/Tables/LegalEntity.sql
@@ -4,7 +4,7 @@
     [Code] NVARCHAR(50) NULL, 
     [DateOfIncorporation] DATETIME NULL,
 	[Status] NVARCHAR(50) NULL,
-	[Source] TINYINT NOT NULL DEFAULT 1,
+	[Source] SMALLINT NOT NULL DEFAULT 1,
 	[PublicSectorDataSource] TINYINT NULL,
 	[Sector] NVARCHAR(100) NULL
 )
